updated code

+++ Bazaar Ghar/Models/WishlistResponse.swift
@@ -499,7 +499,7 @@ struct WishlistMarket: Codable {
 
 // MARK: - Wallet
 struct WishlistWallet: Codable {
-    let balance: Int?
+    let balance: Float?
 }
 
 // MARK: - Variant
